@@ -1,23 +1,22 @@
 name: dokit
 description: 开发工具集
-version: 0.6.0
+version: 0.8.0-nullsafety.0
 homepage: https://www.dokit.cn/#/index/home
 
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  date_format: ^1.0.8
-  vm_service: ^5.5.0
-  package_info: ^0.4.3+2
-  shared_preferences: ^0.5.12+4
-  isolate: ^2.0.3
-  image: ^2.1.19
-  google_fonts: ^1.1.1
+  date_format: ^2.0.2
+  vm_service: ^6.2.0
+  package_info: ^2.0.0
+  shared_preferences: ^2.0.5
+  image: ^3.0.2
+  google_fonts: ^2.0.0
   string_scanner: ^1.0.5
 
 dev_dependencies:
