@@ -9,12 +9,12 @@ import 'package:dokit/kit/apm/http_kit.dart';
 class DoKitHttpOverrides extends HttpOverrides {
   DoKitHttpOverrides(this.origin);
 
-  final HttpOverrides origin;
+  final HttpOverrides? origin;
 
   @override
-  HttpClient createHttpClient(SecurityContext context) {
+  HttpClient createHttpClient(SecurityContext? context) {
     if (origin != null) {
-      return DoKitHttpClient(origin.createHttpClient(context));
+      return DoKitHttpClient(origin!.createHttpClient(context));
     }
     HttpOverrides.global = null;
     final HttpClient client = DoKitHttpClient(HttpClient(context: context));
@@ -27,7 +27,7 @@ class DoKitHttpClient implements HttpClient {
   DoKitHttpClient(this.origin);
 
   final HttpClient origin;
-  HttpInfo httpInfo;
+  HttpInfo? httpInfo;
 
   @override
   set autoUncompress(bool value) => origin.autoUncompress = value;
@@ -42,22 +42,22 @@ class DoKitHttpClient implements HttpClient {
   Duration get idleTimeout => origin.idleTimeout;
 
   @override
-  set connectionTimeout(Duration value) => origin.connectionTimeout = value;
+  set connectionTimeout(Duration? value) => origin.connectionTimeout = value;
 
   @override
-  Duration get connectionTimeout => origin.connectionTimeout;
+  Duration? get connectionTimeout => origin.connectionTimeout;
 
   @override
-  set maxConnectionsPerHost(int value) => origin.maxConnectionsPerHost = value;
+  set maxConnectionsPerHost(int? value) => origin.maxConnectionsPerHost = value;
 
   @override
-  int get maxConnectionsPerHost => origin.maxConnectionsPerHost;
+  int? get maxConnectionsPerHost => origin.maxConnectionsPerHost;
 
   @override
-  set userAgent(String value) => origin.userAgent = value;
+  set userAgent(String? value) => origin.userAgent = value;
 
   @override
-  String get userAgent => origin.userAgent;
+  String? get userAgent => origin.userAgent;
 
   @override
   void addCredentials(
@@ -73,20 +73,20 @@ class DoKitHttpClient implements HttpClient {
 
   @override
   set authenticate(
-      Future<bool> Function(Uri url, String scheme, String realm) f) {
+      Future<bool> Function(Uri url, String scheme, String realm)? f) {
     origin.authenticate = f;
   }
 
   @override
   set authenticateProxy(
-      Future<bool> Function(String host, int port, String scheme, String realm)
+      Future<bool> Function(String host, int port, String scheme, String realm)?
           f) {
     origin.authenticateProxy = f;
   }
 
   @override
   set badCertificateCallback(
-      bool Function(X509Certificate cert, String host, int port) callback) {
+      bool Function(X509Certificate cert, String host, int port)? callback) {
     origin.badCertificateCallback = callback;
   }
 
@@ -96,22 +96,22 @@ class DoKitHttpClient implements HttpClient {
   }
 
   @override
-  set findProxy(String Function(Uri url) f) {
+  set findProxy(String Function(Uri url)? f) {
     origin.findProxy = f;
   }
 
   Future<HttpClientRequest> monitor(Future<HttpClientRequest> future) async {
-    future.catchError((dynamic error, [StackTrace stackTrace]) {
+    future = future.catchError((dynamic error, [StackTrace? stackTrace]) {
       if (httpInfo == null) {
         httpInfo = HttpInfo.error(error.toString());
-        final HttpKit kit = ApmKitManager.instance.getKit(ApmKitName.KIT_HTTP);
-        kit.save(httpInfo);
+        final HttpKit? kit = ApmKitManager.instance.getKit(ApmKitName.KIT_HTTP);
+        kit?.save(httpInfo);
       }
     });
     final HttpClientRequest request = await future;
     httpInfo ??= HttpInfo(request.uri, request.method);
-    final HttpKit kit = ApmKitManager.instance.getKit(ApmKitName.KIT_HTTP);
-    kit.save(httpInfo);
+    final HttpKit? kit = ApmKitManager.instance.getKit(ApmKitName.KIT_HTTP);
+    kit?.save(httpInfo);
     return DoKitHttpClientRequest(request, httpInfo);
   }
 
@@ -195,7 +195,7 @@ class DoKitHttpClientRequest implements HttpClientRequest {
   DoKitHttpClientRequest(this.origin, this.httpInfo);
 
   final HttpClientRequest origin;
-  final HttpInfo httpInfo;
+  final HttpInfo? httpInfo;
 
   @override
   bool get bufferOutput => origin.bufferOutput;
@@ -243,7 +243,7 @@ class DoKitHttpClientRequest implements HttpClientRequest {
   Uri get uri => origin.uri;
 
   @override
-  HttpConnectionInfo get connectionInfo => origin.connectionInfo;
+  HttpConnectionInfo? get connectionInfo => origin.connectionInfo;
 
   @override
   List<Cookie> get cookies => origin.cookies;
@@ -252,7 +252,7 @@ class DoKitHttpClientRequest implements HttpClientRequest {
   Future<HttpClientResponse> get done => origin.done;
 
   @override
-  void write(Object obj) {
+  void write(Object? obj) {
     origin.write(obj);
   }
 
@@ -267,7 +267,7 @@ class DoKitHttpClientRequest implements HttpClientRequest {
   }
 
   @override
-  void writeln([Object obj = '']) {
+  void writeln([dynamic obj = '']) {
     origin.writeln(obj);
   }
 
@@ -279,17 +279,15 @@ class DoKitHttpClientRequest implements HttpClientRequest {
 
   void recordParameter(List<int> data) {
     try {
-      httpInfo.request.header = headers?.toString();
-      if (encoding != null) {
-        httpInfo.request.add(encoding.decode(data));
-      }
+      httpInfo?.request.header = headers.toString();
+      httpInfo?.request.add(encoding.decode(data));
     } catch (e) {
       print(e);
     }
   }
 
   @override
-  void addError(Object error, [StackTrace stackTrace]) {
+  void addError(Object error, [StackTrace? stackTrace]) {
     origin.addError(error, stackTrace);
   }
 
@@ -314,7 +312,7 @@ class DoKitHttpClientRequest implements HttpClientRequest {
   }
 
   void recordResponse(int code, String result, String header, int size) {
-    httpInfo.response.update(code, result, header, size);
+    httpInfo?.response.update(code, result, header, size);
   }
 
   @override
@@ -323,13 +321,13 @@ class DoKitHttpClientRequest implements HttpClientRequest {
   }
 
   @override
-  void abort([Object exception, StackTrace stackTrace]) {
+  void abort([Object? exception, StackTrace? stackTrace]) {
     return origin.abort(exception, stackTrace);
   }
 }
 
 extension HttpClientRequestExt on HttpClientRequest {
-  void abort([Object exception, StackTrace stackTrace]) {
+  void abort([Object? exception, StackTrace? stackTrace]) {
     this.abort(exception, stackTrace);
   }
 }
@@ -347,13 +345,13 @@ class DoKitHttpClientResponse implements HttpClientResponse {
 
   @override
   Stream<List<int>> asBroadcastStream(
-      {void Function(StreamSubscription<List<int>> subscription) onListen,
-      void Function(StreamSubscription<List<int>> subscription) onCancel}) {
+      {void Function(StreamSubscription<List<int>> subscription)? onListen,
+      void Function(StreamSubscription<List<int>> subscription)? onCancel}) {
     return origin.asBroadcastStream(onListen: onListen, onCancel: onCancel);
   }
 
   @override
-  Stream<E> asyncExpand<E>(Stream<E> Function(List<int> event) convert) {
+  Stream<E> asyncExpand<E>(Stream<E>? Function(List<int> event) convert) {
     return asyncExpand(convert);
   }
 
@@ -368,17 +366,17 @@ class DoKitHttpClientResponse implements HttpClientResponse {
   }
 
   @override
-  X509Certificate get certificate => origin.certificate;
+  X509Certificate? get certificate => origin.certificate;
 
   @override
   HttpClientResponseCompressionState get compressionState =>
       origin.compressionState;
 
   @override
-  HttpConnectionInfo get connectionInfo => origin.connectionInfo;
+  HttpConnectionInfo? get connectionInfo => origin.connectionInfo;
 
   @override
-  Future<bool> contains(Object needle) {
+  Future<bool> contains(dynamic needle) {
     return origin.contains(needle);
   }
 
@@ -395,12 +393,12 @@ class DoKitHttpClientResponse implements HttpClientResponse {
 
   @override
   Stream<List<int>> distinct(
-      [bool Function(List<int> previous, List<int> next) equals]) {
+      [bool Function(List<int> previous, List<int> next)? equals]) {
     return origin.distinct(equals);
   }
 
   @override
-  Future<E> drain<E>([E futureValue]) {
+  Future<E> drain<E>([E? futureValue]) {
     return origin.drain(futureValue);
   }
 
@@ -424,7 +422,7 @@ class DoKitHttpClientResponse implements HttpClientResponse {
 
   @override
   Future<List<int>> firstWhere(bool Function(List<int> element) test,
-      {List<int> Function() orElse}) {
+      {List<int> Function()? orElse}) {
     return origin.firstWhere(test, orElse: orElse);
   }
 
@@ -441,7 +439,7 @@ class DoKitHttpClientResponse implements HttpClientResponse {
 
   @override
   Stream<List<int>> handleError(Function onError,
-      {bool Function(dynamic error) test}) {
+      {bool Function(dynamic error)? test}) {
     return origin.handleError(onError, test: test);
   }
 
@@ -467,7 +465,7 @@ class DoKitHttpClientResponse implements HttpClientResponse {
 
   @override
   Future<List<int>> lastWhere(bool Function(List<int> element) test,
-      {List<int> Function() orElse}) {
+      {List<int> Function()? orElse}) {
     return origin.lastWhere(test, orElse: orElse);
   }
 
@@ -475,47 +473,45 @@ class DoKitHttpClientResponse implements HttpClientResponse {
   Future<int> get length => origin.length;
 
   bool isTextResponse() {
-    return headers != null &&
-        headers['content-type'] != null &&
+    return headers['content-type'] != null &&
         (headers['content-type'].toString().contains('json') ||
             headers['content-type'].toString().contains('text') ||
             headers['content-type'].toString().contains('xml'));
   }
 
-  Encoding getEncoding() {
+  Encoding? getEncoding() {
     String charset;
-    if (headers != null &&
-        headers.contentType != null &&
-        headers.contentType.charset != null) {
-      charset = headers.contentType.charset;
+    if (headers.contentType != null && headers.contentType?.charset != null) {
+      charset = headers.contentType!.charset!;
     } else {
       charset = 'utf-8';
     }
-    return Encoding.getByName(charset);
+    return Encoding?.getByName(charset);
   }
 
   @override
-  StreamSubscription<List<int>> listen(void Function(List<int> event) onData,
-      {Function onError, void Function() onDone, bool cancelOnError}) {
+  StreamSubscription<List<int>> listen(void Function(List<int> event)? onData,
+      {Function? onError, void Function()? onDone, bool? cancelOnError}) {
     if (!isTextResponse()) {
       recordResponse(
-          statusCode, '返回结果不支持解析', headers?.toString(), contentLength);
+          statusCode, '返回结果不支持解析', headers.toString(), contentLength);
       return origin.listen(onData,
           onError: onError, onDone: onDone, cancelOnError: cancelOnError);
     }
     void onDataWrapper(List<int> result) {
-      onData(result);
+      onData?.call(result);
       try {
-        if (getEncoding() != null) {
-          recordResponse(statusCode, getEncoding().decode(result),
-              headers?.toString(), contentLength);
+        final encoding = getEncoding();
+        if (encoding != null) {
+          recordResponse(statusCode, encoding.decode(result),
+              headers.toString(), contentLength);
         } else {
           recordResponse(
-              statusCode, '返回结果解析失败', headers?.toString(), contentLength);
+              statusCode, '返回结果解析失败', headers.toString(), contentLength);
         }
       } catch (e) {
         recordResponse(
-            statusCode, '返回结果解析失败', headers?.toString(), contentLength);
+            statusCode, '返回结果解析失败', headers.toString(), contentLength);
       }
     }
 
@@ -541,7 +537,7 @@ class DoKitHttpClientResponse implements HttpClientResponse {
 
   @override
   Future<HttpClientResponse> redirect(
-      [String method, Uri url, bool followLoops]) {
+      [String? method, Uri? url, bool? followLoops]) {
     return origin.redirect(method, url, followLoops);
   }
 
@@ -559,7 +555,7 @@ class DoKitHttpClientResponse implements HttpClientResponse {
 
   @override
   Future<List<int>> singleWhere(bool Function(List<int> element) test,
-      {List<int> Function() orElse}) {
+      {List<int> Function()? orElse}) {
     return origin.singleWhere(test, orElse: orElse);
   }
 
@@ -588,7 +584,7 @@ class DoKitHttpClientResponse implements HttpClientResponse {
 
   @override
   Stream<List<int>> timeout(Duration timeLimit,
-      {void Function(EventSink<List<int>> sink) onTimeout}) {
+      {void Function(EventSink<List<int>> sink)? onTimeout}) {
     return origin.timeout(timeLimit, onTimeout: onTimeout);
   }
 
@@ -607,25 +603,26 @@ class DoKitHttpClientResponse implements HttpClientResponse {
     Stream<S> s = origin.transform<S>(streamTransformer);
     if (!isTextResponse()) {
       recordResponse(
-          statusCode, '返回结果不支持解析', headers?.toString(), contentLength);
+          statusCode, '返回结果不支持解析', headers.toString(), contentLength);
       return s;
     }
     s = s.asBroadcastStream();
     s.listen((S event) {
       if (event is Uint8List) {
         final Uint8List result = event;
-        if (getEncoding() != null) {
-          recordResponse(statusCode, getEncoding().decode(result.toList()),
-              headers?.toString(), event.length);
+        var encoding = getEncoding();
+        if (encoding != null) {
+          recordResponse(statusCode, encoding.decode(result.toList()),
+              headers.toString(), event.length);
         } else {
           recordResponse(
-              statusCode, '返回结果解析失败', headers?.toString(), contentLength);
+              statusCode, '返回结果解析失败', headers.toString(), contentLength);
         }
       } else if (event is String) {
-        recordResponse(statusCode, event, headers?.toString(), contentLength);
+        recordResponse(statusCode, event, headers.toString(), contentLength);
       } else {
         recordResponse(statusCode, 'unknown type:${event.runtimeType}',
-            headers?.toString(), contentLength);
+            headers.toString(), contentLength);
       }
     });
     return s;
