@@ -1,15 +1,14 @@
 import 'package:dokit/kit/apm/fps_kit.dart';
+import 'package:dokit/kit/apm/http_kit.dart';
 import 'package:dokit/kit/apm/launch/page_launch_kit.dart';
 import 'package:dokit/kit/apm/log_kit.dart';
 import 'package:dokit/kit/apm/memory_kit.dart';
 import 'package:dokit/kit/apm/method_channel_kit.dart';
 import 'package:dokit/kit/apm/route_kit.dart';
 import 'package:dokit/kit/apm/source_code_kit.dart';
+import 'package:dokit/kit/kit.dart';
 import 'package:dokit/ui/resident_page.dart';
 import 'package:flutter/material.dart';
-
-import '../kit.dart';
-import 'http_kit.dart';
 
 class ApmKitManager {
   Map<String, ApmKit> kitMap = {
@@ -23,7 +22,7 @@ class ApmKitManager {
     ApmKitName.KIT_PAGE_LAUNCH: PageLaunchKit()
   };
 
-  ApmKitManager._privateConstructor() {}
+  ApmKitManager._privateConstructor();
 
   static final ApmKitManager _instance = ApmKitManager._privateConstructor();
 
@@ -31,12 +30,10 @@ class ApmKitManager {
 
   // 如果想要自定义实现，可以用这个方式进行覆盖。后续扩展入口
   void addKit(String tag, ApmKit kit) {
-    assert(tag != null && kit != null);
     kitMap[tag] = kit;
   }
 
-  T getKit<T extends ApmKit>(String name) {
-    assert(name != null);
+  T? getKit<T extends ApmKit>(String name) {
     if (kitMap.containsKey(name)) {
       return kitMap[name] as T;
     }
@@ -51,7 +48,7 @@ class ApmKitManager {
 }
 
 abstract class ApmKit implements IKit {
-  IStorage storage;
+  late IStorage storage;
 
   void start();
 
@@ -63,22 +60,18 @@ abstract class ApmKit implements IKit {
 
   ApmKit() {
     storage = createStorage();
-    assert(storage != null, 'storage should not be null');
   }
 
   @override
   void tabAction() {
-    ResidentPage.residentPageKey.currentState.setState(() {
+    // ignore: invalid_use_of_protected_member
+    ResidentPage.residentPageKey.currentState?.setState(() {
       ResidentPage.tag = getKitName();
     });
   }
 
-  @override
-  bool save(IInfo info) {
-    return info != null &&
-        storage != null &&
-        !storage.contains(info) &&
-        storage.save(info);
+  bool save(IInfo? info) {
+    return info != null && !storage.contains(info) && storage.save(info);
   }
 
   IStorage getStorage() {
