@@ -1,8 +1,8 @@
 
 class LaunchInfo{
   final int costTime;
-  final String previousPage;
-  final String newPage;
+  final String? previousPage;
+  final String? newPage;
 
   LaunchInfo(this.costTime, this.previousPage, this.newPage);
 }
