@@ -1,15 +1,14 @@
 import 'dart:ui';
 
+import 'package:dokit/dokit.dart';
+import 'package:dokit/kit/apm/apm.dart';
+import 'package:dokit/kit/kit.dart';
 import 'package:dokit/util/time_util.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/src/widgets/framework.dart';
-
-import '../../dokit.dart';
-import '../kit.dart';
-import 'apm.dart';
 
 class ChannelInfo implements IInfo {
   static const int TYPE_USER_SEND = 0;
@@ -18,7 +17,7 @@ class ChannelInfo implements IInfo {
   static const int TYPE_SYSTEM_RECEIVE = 3;
   final String channelName;
 
-  final String method;
+  final String? method;
 
   final dynamic arguments;
   final int startTimestamp;
@@ -27,11 +26,11 @@ class ChannelInfo implements IInfo {
   final int type;
   dynamic results;
   bool expand = false;
-  MethodCodec methodCodec;
-  MessageCodec messageCodec;
+  MethodCodec? methodCodec;
+  MessageCodec? messageCodec;
 
   ChannelInfo(this.channelName, this.method, this.arguments, this.type)
-      : this.startTimestamp = new DateTime.now().millisecondsSinceEpoch;
+      : this.startTimestamp = DateTime.now().millisecondsSinceEpoch;
 
   @override
   String getValue() {
@@ -46,7 +45,6 @@ class ChannelInfo implements IInfo {
         'method:${method}\n' +
         'arguments:${arguments}\n' +
         'results:${results}';
-    ;
   }
 
   factory ChannelInfo.error(String channelName, int type) {
@@ -55,7 +53,7 @@ class ChannelInfo implements IInfo {
 }
 
 class MethodChannelKit extends ApmKit {
-  Function listener;
+  Function? listener;
 
   @override
   Widget createDisplayPage() {
@@ -78,10 +76,10 @@ class MethodChannelKit extends ApmKit {
   }
 
   @override
-  bool save(IInfo info) {
+  bool save(IInfo? info) {
     if (!ChannelPageState.showSystemChannel &&
         ((info as ChannelInfo).type == ChannelInfo.TYPE_SYSTEM_RECEIVE ||
-            (info as ChannelInfo).type == ChannelInfo.TYPE_SYSTEM_SEND)) {
+            info.type == ChannelInfo.TYPE_SYSTEM_SEND)) {
       super.save(info);
       return false;
     }
@@ -116,14 +114,14 @@ class ChannelPage extends StatefulWidget {
 }
 
 class ChannelPageState extends State<ChannelPage> {
-  ScrollController _offsetController =
-      ScrollController(); //定义ListView的controller
+  // 定义ListView的controller
+  ScrollController _offsetController = ScrollController();
   static bool showSystemChannel = false;
 
   Future<void> _listener() async {
     if (!mounted) return;
-    if (SchedulerBinding.instance.schedulerPhase != SchedulerPhase.idle) {
-      await SchedulerBinding.instance.endOfFrame;
+    if (SchedulerBinding.instance?.schedulerPhase != SchedulerPhase.idle) {
+      await SchedulerBinding.instance?.endOfFrame;
       if (!mounted) return;
     }
     setState(() {
@@ -139,7 +137,7 @@ class ChannelPageState extends State<ChannelPage> {
     super.initState();
     ApmKitManager.instance
         .getKit<MethodChannelKit>(ApmKitName.KIT_CHANNEL)
-        .registerListener(_listener);
+        ?.registerListener(_listener);
   }
 
   @override
@@ -147,20 +145,21 @@ class ChannelPageState extends State<ChannelPage> {
     super.dispose();
     ApmKitManager.instance
         .getKit<MethodChannelKit>(ApmKitName.KIT_CHANNEL)
-        .unregisterListener();
+        ?.unregisterListener();
+    _offsetController.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    List<IInfo> items = ApmKitManager.instance
+    List<IInfo>? items = ApmKitManager.instance
         .getKit<MethodChannelKit>(ApmKitName.KIT_CHANNEL)
-        .getStorage()
+        ?.getStorage()
         .getAll()
         .reversed
         .where((element) => showSystemChannel
             ? true
             : ((element as ChannelInfo).type == ChannelInfo.TYPE_USER_SEND ||
-                (element as ChannelInfo).type == ChannelInfo.TYPE_USER_RECEIVE))
+                (element).type == ChannelInfo.TYPE_USER_RECEIVE))
         .toList();
     return Column(
       children: <Widget>[
@@ -209,16 +208,14 @@ class ChannelPageState extends State<ChannelPage> {
                       this.setState(() {
                         ApmKitManager.instance
                             .getKit<MethodChannelKit>(ApmKitName.KIT_CHANNEL)
-                            .getStorage()
+                            ?.getStorage()
                             .clear();
                       });
                     },
                     child: Container(
-                      decoration: new BoxDecoration(
-                        border:
-                            new Border.all(color: Color(0xff337cc4), width: 1),
-                        borderRadius:
-                            new BorderRadius.circular(2), // 也可控件一边圆角大小
+                      decoration: BoxDecoration(
+                        border: Border.all(color: Color(0xff337cc4), width: 1),
+                        borderRadius: BorderRadius.circular(2), // 也可控件一边圆角大小
                       ),
                       margin: EdgeInsets.only(left: 10),
                       padding: EdgeInsets.all(2),
@@ -235,11 +232,9 @@ class ChannelPageState extends State<ChannelPage> {
                       _offsetController.jumpTo(0);
                     },
                     child: Container(
-                      decoration: new BoxDecoration(
-                        border:
-                            new Border.all(color: Color(0xff337cc4), width: 1),
-                        borderRadius:
-                            new BorderRadius.circular(2), // 也可控件一边圆角大小
+                      decoration: BoxDecoration(
+                        border: Border.all(color: Color(0xff337cc4), width: 1),
+                        borderRadius: BorderRadius.circular(2), // 也可控件一边圆角大小
                       ),
                       margin: EdgeInsets.only(left: 10),
                       padding: EdgeInsets.all(2),
@@ -256,20 +251,22 @@ class ChannelPageState extends State<ChannelPage> {
           child: Container(
               alignment: Alignment.topCenter,
               color: Color(0xfff5f6f7),
-              child: ListView.builder(
-                  controller: _offsetController,
-                  itemCount: items.length,
-                  padding:
-                      EdgeInsets.only(left: 4, right: 4, bottom: 0, top: 0),
-                  reverse: true,
-                  shrinkWrap: true,
-                  itemBuilder: (context, index) {
-                    return ChannelItemWidget(
-                      item: items[index],
-                      index: index,
-                      isLast: index == items.length - 1,
-                    );
-                  })),
+              child: items == null
+                  ? null
+                  : ListView.builder(
+                      controller: _offsetController,
+                      itemCount: items.length,
+                      padding:
+                          EdgeInsets.only(left: 4, right: 4, bottom: 0, top: 0),
+                      reverse: true,
+                      shrinkWrap: true,
+                      itemBuilder: (context, index) {
+                        return ChannelItemWidget(
+                          item: items[index] as ChannelInfo,
+                          index: index,
+                          isLast: index == items.length - 1,
+                        );
+                      })),
         )
       ],
     );
@@ -282,10 +279,7 @@ class ChannelItemWidget extends StatefulWidget {
   final bool isLast;
 
   ChannelItemWidget(
-      {Key key,
-      @required this.item,
-      @required this.index,
-      @required this.isLast})
+      {Key? key, required this.item, required this.index, required this.isLast})
       : super(key: key);
 
   @override
@@ -313,114 +307,117 @@ class _ChannelItemWidgetState extends State<ChannelItemWidget> {
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-        onTap: () {
-          setState(() {
-            widget.item.expand = !widget.item.expand;
-          });
-        },
-        child: Card(
-            color: Colors.white,
-            child: Row(mainAxisAlignment: MainAxisAlignment.start, children: <
-                Widget>[
-              Container(
-                  width: MediaQuery.of(context).size.width - 80,
-                  margin:
-                      EdgeInsets.only(left: 16, right: 16, top: 12, bottom: 12),
-                  child: RichText(
-                      maxLines: widget.item.expand ? 9999 : 7,
-                      overflow: TextOverflow.ellipsis,
-                      text: TextSpan(children: [
-                        TextSpan(
-                            text:
-                                '[${toTimeString(widget.item.startTimestamp)}]',
-                            style: TextStyle(
-                                fontSize: 9,
-                                color: Color(0xff333333),
-                                height: 1.2)),
-                        WidgetSpan(
-                            child: Container(
-                                child: Text('${getChannelType()}',
-                                    style: TextStyle(
-                                        fontSize: 8,
-                                        color: Color(0xffffffff),
-                                        height: 1.2)),
-                                height: 11,
-                                margin: EdgeInsets.only(left: 4),
-                                padding: EdgeInsets.only(left: 6, right: 6),
-                                decoration: BoxDecoration(
-                                    borderRadius:
-                                        BorderRadius.all(Radius.circular(2)),
-                                    color: (widget.item.type % 2 != 0
-                                        ? Color(0xffd0607e)
-                                        : Color(0xff337cc4))))),
-                        TextSpan(
-                            text:
-                                '  Cost:${widget.item.endTimestamp > 0 ? ((widget.item.endTimestamp - widget.item.startTimestamp).toString() + 'ms') : '-'} ',
-                            style: TextStyle(
+      onTap: () {
+        setState(() {
+          widget.item.expand = !widget.item.expand;
+        });
+      },
+      child: Card(
+        color: Colors.white,
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.start,
+          children: <Widget>[
+            Container(
+                width: MediaQuery.of(context).size.width - 80,
+                margin:
+                    EdgeInsets.only(left: 16, right: 16, top: 12, bottom: 12),
+                child: RichText(
+                    maxLines: widget.item.expand ? 9999 : 7,
+                    overflow: TextOverflow.ellipsis,
+                    text: TextSpan(children: [
+                      TextSpan(
+                          text: '[${toTimeString(widget.item.startTimestamp)}]',
+                          style: TextStyle(
                               fontSize: 9,
-                              color: Color(0xff666666),
+                              color: Color(0xff333333),
+                              height: 1.2)),
+                      WidgetSpan(
+                          child: Container(
+                              child: Text('${getChannelType()}',
+                                  style: TextStyle(
+                                      fontSize: 8,
+                                      color: Color(0xffffffff),
+                                      height: 1.2)),
+                              height: 11,
+                              margin: EdgeInsets.only(left: 4),
+                              padding: EdgeInsets.only(left: 6, right: 6),
+                              decoration: BoxDecoration(
+                                  borderRadius:
+                                      BorderRadius.all(Radius.circular(2)),
+                                  color: (widget.item.type % 2 != 0
+                                      ? Color(0xffd0607e)
+                                      : Color(0xff337cc4))))),
+                      TextSpan(
+                          text:
+                              '  Cost:${widget.item.endTimestamp > 0 ? ((widget.item.endTimestamp - widget.item.startTimestamp).toString() + 'ms') : '-'} ',
+                          style: TextStyle(
+                            fontSize: 9,
+                            color: Color(0xff666666),
+                            height: 1.5,
+                          )),
+                      TextSpan(
+                          text: '\nChannelName: ',
+                          style: TextStyle(
+                              fontSize: 10,
+                              color: Color(0xff333333),
                               height: 1.5,
-                            )),
-                        TextSpan(
-                            text: '\nChannelName: ',
-                            style: TextStyle(
-                                fontSize: 10,
-                                color: Color(0xff333333),
-                                height: 1.5,
-                                fontWeight: FontWeight.bold)),
-                        TextSpan(
-                            text: '${widget.item.channelName}',
-                            style: TextStyle(
-                                fontSize: 10,
-                                height: 1.5,
-                                color: Color(0xff666666))),
-                        TextSpan(
-                            text: '\nMethod: ',
-                            style: TextStyle(
-                                height: 1.5,
-                                fontSize: 10,
-                                color: Color(0xff333333),
-                                fontWeight: FontWeight.bold)),
-                        TextSpan(
-                            text: '${widget.item.method}',
-                            style: TextStyle(
-                                fontSize: 10,
-                                height: 1.5,
-                                color: Color(0xff666666))),
-                        TextSpan(
-                            text: '\nArguments: ',
-                            style: TextStyle(
-                                fontSize: 10,
-                                height: 1.5,
-                                color: Color(0xff333333),
-                                fontWeight: FontWeight.bold)),
-                        TextSpan(
-                            text: '${widget.item.arguments}',
-                            style: TextStyle(
-                                fontSize: 10,
-                                height: 1.5,
-                                color: Color(0xff666666))),
-                        TextSpan(
-                            text: '\nResult: ',
-                            style: TextStyle(
-                                fontSize: 10,
-                                height: 1.5,
-                                color: Color(0xff333333),
-                                fontWeight: FontWeight.bold)),
-                        TextSpan(
-                            text: '${widget.item.results}',
-                            style: TextStyle(
-                                fontSize: 10,
-                                height: 1.5,
-                                color: Color(0xff666666))),
-                      ]))),
-              Image.asset(
-                  widget.item.expand
-                      ? 'images/dk_channel_expand_h.png'
-                      : 'images/dk_channel_expand_n.png',
-                  package: DK_PACKAGE_NAME,
-                  height: 14,
-                  width: 9)
-            ])));
+                              fontWeight: FontWeight.bold)),
+                      TextSpan(
+                          text: '${widget.item.channelName}',
+                          style: TextStyle(
+                              fontSize: 10,
+                              height: 1.5,
+                              color: Color(0xff666666))),
+                      TextSpan(
+                          text: '\nMethod: ',
+                          style: TextStyle(
+                              height: 1.5,
+                              fontSize: 10,
+                              color: Color(0xff333333),
+                              fontWeight: FontWeight.bold)),
+                      TextSpan(
+                          text: '${widget.item.method}',
+                          style: TextStyle(
+                              fontSize: 10,
+                              height: 1.5,
+                              color: Color(0xff666666))),
+                      TextSpan(
+                          text: '\nArguments: ',
+                          style: TextStyle(
+                              fontSize: 10,
+                              height: 1.5,
+                              color: Color(0xff333333),
+                              fontWeight: FontWeight.bold)),
+                      TextSpan(
+                          text: '${widget.item.arguments}',
+                          style: TextStyle(
+                              fontSize: 10,
+                              height: 1.5,
+                              color: Color(0xff666666))),
+                      TextSpan(
+                          text: '\nResult: ',
+                          style: TextStyle(
+                              fontSize: 10,
+                              height: 1.5,
+                              color: Color(0xff333333),
+                              fontWeight: FontWeight.bold)),
+                      TextSpan(
+                          text: '${widget.item.results}',
+                          style: TextStyle(
+                              fontSize: 10,
+                              height: 1.5,
+                              color: Color(0xff666666))),
+                    ]))),
+            Image.asset(
+                widget.item.expand
+                    ? 'images/dk_channel_expand_h.png'
+                    : 'images/dk_channel_expand_n.png',
+                package: DK_PACKAGE_NAME,
+                height: 14,
+                width: 9)
+          ],
+        ),
+      ),
+    );
   }
 }
