@@ -1,7 +1,6 @@
 import 'dart:developer';
 
 import 'package:dokit/kit/apm/vm/version.dart';
-import 'package:flutter/foundation.dart';
 import 'package:vm_service/utils.dart';
 import 'package:vm_service/vm_service.dart';
 import 'package:vm_service/vm_service_io.dart';
@@ -14,13 +13,13 @@ class VMServiceWrapper {
 
   static VMServiceWrapper get instance => _instance;
 
-  VmService service;
+  VmService? service;
 
-  IsolateRef main;
+  IsolateRef? main;
 
-  VM vm;
+  VM? vm;
 
-  ExtensionService _extensionService;
+  ExtensionService? _extensionService;
   bool connected = false;
 
   Future<VmService> getService(info) async {
@@ -30,39 +29,40 @@ class VMServiceWrapper {
 
   Future<void> connect() async {
     ServiceProtocolInfo info = await Service.getInfo();
-    if (info == null || info.serverUri == null) {
+    if (info.serverUri == null) {
       print("service  protocol url is null,start vm service fail");
       return;
     }
     service = await getService(info);
     print('socket connected in service $info');
-    vm = await service.getVM();
-    List<IsolateRef> isolates = vm.isolates;
-    main = isolates.firstWhere((ref) => ref.name.contains('main'));
-    main ??= isolates.first;
+    vm = await service?.getVM();
+    List<IsolateRef>? isolates = vm?.isolates;
+    main = isolates?.firstWhere((ref) => ref.name?.contains('main') == true);
+    main ??= isolates?.first;
     connected = true;
   }
 
-  Future<Response> callExtensionService(String method) async {
-    if (_extensionService == null) {
-      _extensionService = new ExtensionService(service, main);
-      await _extensionService.loadExtensionService();
+  Future<Response?> callExtensionService(String method) async {
+    if (_extensionService == null && service != null && main != null) {
+      _extensionService = ExtensionService(service!, main!);
+      await _extensionService?.loadExtensionService();
     }
-    return _extensionService.callMethod(method);
+    return _extensionService!.callMethod(method);
   }
 
   gc() {
-    if (connected) {
-      service.getAllocationProfile(main.id, gc: true);
+    var isolateId = main?.id;
+    if (connected && isolateId != null) {
+      service?.getAllocationProfile(isolateId, gc: true);
     }
   }
 
   disConnect() async {
     if (service != null) {
       print('waiting for client to shut down...');
-      service.dispose();
+      await service?.dispose();
 
-      await service.onDone;
+      await service?.onDone;
       connected = false;
       service = null;
       print('service client shut down');
@@ -73,8 +73,8 @@ class VMServiceWrapper {
 class ExtensionService {
   final VmService serviceClient;
   final IsolateRef main;
-  Version _protocolVersion;
-  Version _dartIoVersion;
+  Version? _protocolVersion;
+  Version? _dartIoVersion;
 
   Map<String, List<String>> get registeredMethodsForService =>
       _registeredMethodsForService;
@@ -115,38 +115,41 @@ class ExtensionService {
           : '_Service';
 
   Future<bool> isProtocolVersionSupported({
-    @required SemanticVersion supportedVersion,
+    required SemanticVersion supportedVersion,
   }) async {
     _protocolVersion ??= await serviceClient.getVersion();
     return isProtocolVersionSupportedNow(supportedVersion: supportedVersion);
   }
 
   bool isProtocolVersionSupportedNow({
-    @required SemanticVersion supportedVersion,
+    required SemanticVersion supportedVersion,
   }) {
+    if (_protocolVersion == null) {
+      return false;
+    }
     return _versionSupported(
-      version: _protocolVersion,
+      version: _protocolVersion!,
       supportedVersion: supportedVersion,
     );
   }
 
   bool _versionSupported({
-    @required Version version,
-    @required SemanticVersion supportedVersion,
+    required Version version,
+    required SemanticVersion supportedVersion,
   }) {
     return SemanticVersion(
-      major: version.major,
-      minor: version.minor,
+      major: version.major ?? 0,
+      minor: version.minor ?? 0,
     ).isSupported(supportedVersion: supportedVersion);
   }
 
   Future<bool> isDartIoVersionSupported({
-    @required SemanticVersion supportedVersion,
-    @required String isolateId,
+    required SemanticVersion supportedVersion,
+    required String isolateId,
   }) async {
     _dartIoVersion ??= await getDartIOVersion(isolateId);
     return _versionSupported(
-      version: _dartIoVersion,
+      version: _dartIoVersion!,
       supportedVersion: supportedVersion,
     );
   }
@@ -155,11 +158,11 @@ class ExtensionService {
       serviceClient.getDartIOVersion(isolateId);
 
   void handleServiceEvent(Event e) {
-    if (e.kind == EventKind.kServiceRegistered) {
-      final serviceName = e.service;
+    if (e.kind == EventKind.kServiceRegistered && e.method != null) {
+      final serviceName = e.service ?? '';
       _registeredMethodsForService
           .putIfAbsent(serviceName, () => [])
-          .add(e.method);
+          .add(e.method!);
     }
 
     if (e.kind == EventKind.kServiceUnregistered) {
@@ -168,12 +171,13 @@ class ExtensionService {
     }
   }
 
-  Future<Response> callMethod(String method) {
+  Future<Response?> callMethod(String method) {
     if (registeredMethodsForService.containsKey(method)) {
-      return (serviceClient.callMethod(registeredMethodsForService[method].last,
+      return (serviceClient.callMethod(
+          registeredMethodsForService[method]!.last,
           isolateId: main.id));
     }
-    return null;
+    return Future.value(null);
   }
 }
 
