@@ -1,3 +1,12 @@
+// Copyright© Dokit for Flutter.
+//
+// color_pick.dart
+// Flutter
+//
+// Created by linusflow on 2021/5/12
+// Modified by linusflow on 2021/5/12 下午2:28
+//
+
 import 'dart:typed_data';
 import 'dart:ui' as ui;
 
@@ -28,26 +37,26 @@ class ColorPickerKit extends VisualKit {
   static final ColorPickerKit _instance = ColorPickerKit._privateConstructor();
   static ColorPickerKit get instance => _instance;
 
-  bool isShown;
+  late bool isShown;
   // 选中的颜色
   final ValueNotifier<Color> color = ValueNotifier<Color>(Colors.white);
 
   // 当前屏幕的截图快照
-  final ValueNotifier<ui.Image> snapshot = ValueNotifier<ui.Image>(null);
+  final ValueNotifier<ui.Image?> snapshot = ValueNotifier<ui.Image?>(null);
 
   // 放大镜当前位置（左上角）
-  ValueNotifier<Offset> _position;
+  late ValueNotifier<Offset> _position;
   ValueNotifier<Offset> get position => _position;
 
   // 放大镜的直径
   final double diameter = 170;
   // 像素点放大的倍数
   final double scale = 8;
-  Offset _initPosition;
+  late Offset _initPosition;
   Offset get initPosition => _initPosition;
 
-  OverlayEntry _focusEntry;
-  OverlayEntry _infoEntry;
+  late OverlayEntry _focusEntry;
+  late OverlayEntry _infoEntry;
 
   @override
   String getIcon() {
@@ -61,22 +70,22 @@ class ColorPickerKit extends VisualKit {
 
   @override
   void tabAction() {
-    final DoKitBtnState state = DoKitBtn.doKitBtnKey.currentState;
-    state.closeDebugPage();
-    show(DoKitBtn.doKitBtnKey.currentContext, state.owner);
+    final DoKitBtnState? state = DoKitBtn.doKitBtnKey.currentState;
+    state?.closeDebugPage();
+    show(DoKitBtn.doKitBtnKey.currentContext, state?.owner);
   }
 
-  static void show(BuildContext context, OverlayEntry entrance) {
+  static void show(BuildContext? context, OverlayEntry? entrance) {
     _instance._show(context, entrance);
   }
 
-  void _show(BuildContext context, OverlayEntry entrance) {
+  void _show(BuildContext? context, OverlayEntry? entrance) {
     if (isShown) {
       return;
     }
     isShown = true;
-    doKitOverlayKey.currentState.insert(_focusEntry, below: entrance);
-    doKitOverlayKey.currentState.insert(_infoEntry, below: _focusEntry);
+    doKitOverlayKey.currentState?.insert(_focusEntry, below: entrance);
+    doKitOverlayKey.currentState?.insert(_infoEntry, below: _focusEntry);
   }
 
   static bool hide(BuildContext context) {
@@ -101,10 +110,10 @@ class ColorPickerWidget extends StatefulWidget {
 
 class ColorPickerWidgetState extends State<ColorPickerWidget> {
   // 当前页面的截图快照
-  images.Image _image;
-  Future<images.Image> get image async {
+  images.Image? _image;
+  Future<images.Image?> get image async {
     if (_image != null) {
-      return _image;
+      return _image!;
     }
     await _updateImage();
 
@@ -114,15 +123,15 @@ class ColorPickerWidgetState extends State<ColorPickerWidget> {
   // 是否做好显示颜色拾取器的准备
   bool _ready = false;
 
-  Uint8List _imageUint8List;
-  Future<Uint8List> get imageUint8List async {
+  Uint8List? _imageUint8List;
+  Future<Uint8List?> get imageUint8List async {
     if (_imageUint8List != null) {
-      return _imageUint8List;
+      return _imageUint8List!;
     }
 
-    final RenderRepaintBoundary boundary =
+    final RenderRepaintBoundary? boundary =
         _findCurrentPageRepaintBoundaryRenderObject();
-    final Uint8List imageData = await _boundaryToImageUint8List(boundary);
+    final Uint8List? imageData = await _boundaryToImageUint8List(boundary);
     _imageUint8List = imageData;
 
     return _imageUint8List;
@@ -133,14 +142,14 @@ class ColorPickerWidgetState extends State<ColorPickerWidget> {
   set position(Offset point) => ColorPickerKit.instance.position.value = point;
 
   // 手指和屏幕接触的初始位置
-  Offset _touchPoint;
+  Offset? _touchPoint;
   // 手指和屏幕接触时，放大镜左上角的位置
   Offset _lastPosition = ColorPickerKit.instance.position.value;
   Offset get deltaOffset {
-    if (_touchPoint == null || _lastPosition == null) {
+    if (_touchPoint == null) {
       return Offset.zero;
     }
-    return _touchPoint - _lastPosition;
+    return _touchPoint! - _lastPosition;
   }
 
   // 放大镜的半径
@@ -149,7 +158,7 @@ class ColorPickerWidgetState extends State<ColorPickerWidget> {
   @override
   initState() {
     super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance?.addPostFrameCallback((_) {
       // 第一次内部会调用_updateImage方法
       _updateColor();
     });
@@ -169,7 +178,7 @@ class ColorPickerWidgetState extends State<ColorPickerWidget> {
           position = event.position - deltaOffset;
           _updateColor();
         },
-        child: Draggable<dynamic>(
+        child: Draggable(
           child: _buildMagnifier(context),
           feedback: _buildMagnifier(context),
           childWhenDragging: Container(),
@@ -225,11 +234,14 @@ class ColorPickerWidgetState extends State<ColorPickerWidget> {
             children: [
               ValueListenableBuilder<Offset>(
                 valueListenable: ColorPickerKit.instance.position,
-                builder: (BuildContext context, Offset value, Widget child) {
-                  return ValueListenableBuilder<ui.Image>(
+                builder: (BuildContext context, Offset value, Widget? child) {
+                  if (ColorPickerKit.instance.snapshot.value == null) {
+                    return Container();
+                  }
+                  return ValueListenableBuilder<ui.Image?>(
                     valueListenable: ColorPickerKit.instance.snapshot,
                     builder:
-                        (BuildContext context, ui.Image value, Widget child) {
+                        (BuildContext context, ui.Image? value, Widget? child) {
                       return CustomPaint(
                         painter: GridsPainter(),
                         size: Size.fromRadius(_radius),
@@ -271,31 +283,37 @@ class ColorPickerWidgetState extends State<ColorPickerWidget> {
   Future<void> _updateColor() async {
     final double dpr = ui.window.devicePixelRatio;
     final center = (position + Offset(_radius, _radius)) * dpr;
-    final images.Image image = await this.image;
-    final abgrPixel = image.getPixelSafe(center.dx.toInt(), center.dy.toInt());
+    final images.Image? image = await this.image;
+    final abgrPixel = image?.getPixelSafe(center.dx.toInt(), center.dy.toInt());
+    if (abgrPixel == null) {
+      return;
+    }
     final int argbPixel = _abgrToArgb(abgrPixel);
     ColorPickerKit.instance.color.value = Color(argbPixel);
   }
 
-  RenderRepaintBoundary _findCurrentPageRepaintBoundaryRenderObject() {
+  RenderRepaintBoundary? _findCurrentPageRepaintBoundaryRenderObject() {
     final owner = context.findRenderObject()?.owner;
     assert(owner != null, '当前正在build，无法获取当前页面的RepaintBoundary！');
 
-    bool isRepaintBoundaryTo_ModalScopeStatus(String desc) {
+    bool isRepaintBoundaryTo_ModalScopeStatus(String? desc) {
       if (desc?.isEmpty ?? true) {
         return false;
       }
-      final creators = desc.split(' ← ');
+      final creators = desc!.split(' ← ');
       const sampleCreators = [
         'RepaintBoundary',
         '_FocusMarker',
         'Semantics',
         'FocusScope',
+        'PrimaryScrollController', // flutter2.0+特有
         '_ActionsMarker',
         'Actions',
+        'Builder', // flutter2.0+特有
         'PageStorage',
         'Offstage',
         '_ModalScopeStatus',
+        'UnmanagedRestorationScope', // flutter2.0+特有
       ];
       for (int i = 0; i < sampleCreators.length; i++) {
         if (creators.length < i + 1) {
@@ -309,20 +327,20 @@ class ColorPickerWidgetState extends State<ColorPickerWidget> {
       return true;
     }
 
-    final ModalRoute<dynamic> rootRoute =
-        ModalRoute.of<dynamic>(DoKitApp.appKey.currentContext);
+    final ModalRoute<dynamic>? rootRoute =
+        ModalRoute.of<dynamic>(DoKitApp.appKey.currentContext!);
 
     // 当前页面的_ModalScopeStatus下的RepaintBoundary的RenderObject
-    RenderRepaintBoundary currentPageRepaintBoundary;
+    RenderRepaintBoundary? currentPageRepaintBoundary;
     void filter(Element element) {
       if (element is RenderObjectElement && element.renderObject is RenderBox) {
-        final ModalRoute<dynamic> route = ModalRoute.of<dynamic>(element);
+        final ModalRoute<dynamic>? route = ModalRoute.of<dynamic>(element);
         if (route != null && route != rootRoute) {
           final RenderBox renderBox = element.renderObject as RenderBox;
           if (renderBox.hasSize &&
               renderBox.attached &&
               renderBox.isRepaintBoundary) {
-            String desc;
+            String? desc;
             if (!kReleaseMode) {
               desc = element.renderObject.debugCreator.toString();
             }
@@ -336,21 +354,21 @@ class ColorPickerWidgetState extends State<ColorPickerWidget> {
       element.visitChildren(filter);
     }
 
-    DoKitApp.appKey.currentContext.visitChildElements(filter);
+    DoKitApp.appKey.currentContext?.visitChildElements(filter);
 
     return currentPageRepaintBoundary;
   }
 
-  Future<Uint8List> _boundaryToImageUint8List(
-      RenderRepaintBoundary boundary) async {
+  Future<Uint8List?> _boundaryToImageUint8List(
+      RenderRepaintBoundary? boundary) async {
     if (boundary == null) {
       return null;
     }
     final double dpr = ui.window.devicePixelRatio;
     final ui.Image image = await boundary.toImage(pixelRatio: dpr);
-    final ByteData byteData =
+    final ByteData? byteData =
         await image.toByteData(format: ui.ImageByteFormat.png);
-    final Uint8List pngBytes = byteData.buffer.asUint8List();
+    final Uint8List? pngBytes = byteData?.buffer.asUint8List();
 
     return pngBytes;
   }
@@ -367,7 +385,7 @@ class GridsPainter extends CustomPainter {
   GridsPainter();
 
   double get scale => ColorPickerKit.instance.scale;
-  ui.Image get image => ColorPickerKit.instance.snapshot.value;
+  ui.Image? get image => ColorPickerKit.instance.snapshot.value;
   Offset get position => ColorPickerKit.instance.position.value;
   double get radius => ColorPickerKit.instance.diameter / 2;
 
@@ -379,7 +397,7 @@ class GridsPainter extends CustomPainter {
 
   @override
   void paint(Canvas canvas, Size size) {
-    if (image == null || position == null) {
+    if (image == null) {
       return;
     }
 
@@ -392,7 +410,7 @@ class GridsPainter extends CustomPainter {
     final srcRect =
         Rect.fromLTWH(left, top, width.toDouble(), height.toDouble());
     final distRect = Rect.fromLTWH(0, 0, size.width, size.height);
-    canvas.drawImageRect(image, srcRect, distRect, paint);
+    canvas.drawImageRect(image!, srcRect, distRect, paint);
   }
 
   @override
@@ -460,16 +478,16 @@ class ColorPickerInfoWidget extends StatefulWidget {
 
 class _ColorPickerInfoWidgetState extends State<ColorPickerInfoWidget> {
   double top = infoWidgetTopMargin;
-  Color color;
+  Color? color;
   String get colorDesc =>
-      '#${color?.value?.toRadixString(16)?.padLeft(8, '0')?.toUpperCase() ?? ''}';
+      '#${color?.value.toRadixString(16).padLeft(8, '0').toUpperCase() ?? ''}';
 
   @override
   Widget build(BuildContext context) {
     return Positioned(
       left: infoWidgetHorizontalMargin,
       top: top,
-      child: Draggable<dynamic>(
+      child: Draggable(
           child: _buildInfoView(),
           feedback: _buildInfoView(),
           childWhenDragging: Container(),
@@ -511,7 +529,7 @@ class _ColorPickerInfoWidgetState extends State<ColorPickerInfoWidget> {
           children: <Widget>[
             ValueListenableBuilder<Color>(
               valueListenable: ColorPickerKit.instance.color,
-              builder: (BuildContext context, Color value, Widget child) {
+              builder: (BuildContext context, Color value, Widget? child) {
                 color = value;
                 return Row(
                   crossAxisAlignment: CrossAxisAlignment.center,
