@@ -30,23 +30,19 @@ class DashedDecoration extends Decoration {
     this.gap = 2.0,
     this.dashedColor,
     this.dawDashed = true,
-  })  : assert(shape != null),
-        assert(
-            backgroundBlendMode == null || color != null || gradient != null,
-            "backgroundBlendMode applies to DashedDecoration's background color or "
-            'gradient, but no color or gradient was provided.');
+  });
 
   /// Creates a copy of this object but with the given fields replaced with the
-  /// new values.
+  /// values.
   DashedDecoration copyWith({
-    Color color,
-    DecorationImage image,
-    BoxBorder border,
-    BorderRadiusGeometry borderRadius,
-    List<BoxShadow> boxShadow,
-    Gradient gradient,
-    BlendMode backgroundBlendMode,
-    BoxShape shape,
+    Color? color,
+    DecorationImage? image,
+    BoxBorder? border,
+    BorderRadiusGeometry? borderRadius,
+    List<BoxShadow>? boxShadow,
+    Gradient? gradient,
+    BlendMode? backgroundBlendMode,
+    BoxShape? shape,
   }) {
     return DashedDecoration(
       color: color ?? this.color,
@@ -57,6 +53,7 @@ class DashedDecoration extends Decoration {
       gradient: gradient ?? this.gradient,
       backgroundBlendMode: backgroundBlendMode ?? this.backgroundBlendMode,
       shape: shape ?? this.shape,
+      dashedColor: this.dashedColor,
     );
   }
 
@@ -75,14 +72,14 @@ class DashedDecoration extends Decoration {
   /// This is ignored if [gradient] is non-null.
   ///
   /// The [color] is drawn under the [image].
-  final Color color;
+  final Color? color;
 
   /// An image to paint above the background [color] or [gradient].
   ///
   /// If [shape] is [BoxShape.circle] then the image is clipped to the circle's
   /// boundary; if [borderRadius] is non-null then the image is clipped to the
   /// given radii.
-  final DecorationImage image;
+  final DecorationImage? image;
 
   /// A border to draw above the background [color], [gradient], or [image].
   ///
@@ -94,7 +91,7 @@ class DashedDecoration extends Decoration {
   /// Use [BoxBorder] objects to describe borders that should flip their left
   /// and right edges based on whether the text is being read left-to-right or
   /// right-to-left.
-  final BoxBorder border;
+  final BoxBorder? border;
 
   /// If non-null, the corners of this box are rounded by this [BorderRadius].
   ///
@@ -102,7 +99,7 @@ class DashedDecoration extends Decoration {
   /// [BoxShape.rectangle].
   ///
   /// {@macro flutter.painting.boxDecoration.clip}
-  final BorderRadiusGeometry borderRadius;
+  final BorderRadiusGeometry? borderRadius;
 
   /// A list of shadows cast by this box behind the box.
   ///
@@ -113,14 +110,14 @@ class DashedDecoration extends Decoration {
   ///  * [kElevationToShadow], for some predefined shadows used in Material
   ///    Design.
   ///  * [PhysicalModel], a widget for showing shadows.
-  final List<BoxShadow> boxShadow;
+  final List<BoxShadow>? boxShadow;
 
   /// A gradient to use when filling the box.
   ///
   /// If this is specified, [color] has no effect.
   ///
   /// The [gradient] is drawn under the [image].
-  final Gradient gradient;
+  final Gradient? gradient;
 
   /// The blend mode applied to the [color] or [gradient] background of the box.
   ///
@@ -128,7 +125,7 @@ class DashedDecoration extends Decoration {
   /// mode is used.
   ///
   /// If no [color] or [gradient] is provided then the blend mode has no impact.
-  final BlendMode backgroundBlendMode;
+  final BlendMode? backgroundBlendMode;
 
   /// The shape to fill the background [color], [gradient], and [image] into and
   /// to cast as the [boxShadow].
@@ -143,36 +140,39 @@ class DashedDecoration extends Decoration {
   /// [BoxShape.rectangle].
   ///
   /// {@macro flutter.painting.boxDecoration.clip}
-  final BoxShape shape;
+  final BoxShape? shape;
 
   final double strokeHeight;
   final double gap;
-  final Color dashedColor;
-  final bool dawDashed;
+  final Color? dashedColor;
+  final bool? dawDashed;
 
   @override
-  EdgeInsetsGeometry get padding => border?.dimensions;
+  EdgeInsetsGeometry? get padding => border?.dimensions;
 
   @override
   Path getClipPath(Rect rect, TextDirection textDirection) {
-    Path clipPath;
+    late Path clipPath;
     switch (shape) {
       case BoxShape.circle:
         clipPath = Path()..addOval(rect);
         break;
       case BoxShape.rectangle:
-        if (borderRadius != null)
+        if (borderRadius != null) {
           clipPath = Path()
-            ..addRRect(borderRadius.resolve(textDirection).toRRect(rect));
+            ..addRRect(borderRadius!.resolve(textDirection).toRRect(rect));
+        }
         break;
+      default:
+        clipPath = Path()..addOval(rect);
     }
     return clipPath;
   }
 
-  /// Returns a new box decoration that is scaled by the given factor.
+  /// Returns a box decoration that is scaled by the given factor.
   DashedDecoration scale(double factor) {
     return DashedDecoration(
-      color: Color.lerp(null, color, factor),
+      color: Color.lerp(null, color, factor)!,
       image: image,
 // TODO(ianh): fade the image from transparent
       border: BoxBorder.lerp(null, border, factor),
@@ -187,7 +187,7 @@ class DashedDecoration extends Decoration {
   bool get isComplex => boxShadow != null;
 
   @override
-  DashedDecoration lerpFrom(Decoration a, double t) {
+  DashedDecoration? lerpFrom(Decoration? a, double t) {
     if (a == null) {
       return scale(t);
     }
@@ -198,7 +198,7 @@ class DashedDecoration extends Decoration {
   }
 
   @override
-  DashedDecoration lerpTo(Decoration b, double t) {
+  DashedDecoration? lerpTo(Decoration? b, double t) {
     if (b == null) {
       return scale(1.0 - t);
     }
@@ -232,14 +232,13 @@ class DashedDecoration extends Decoration {
   ///  * [lerpFrom] and [lerpTo], which are used to implement [Decoration.lerp]
   ///    and which use [DashedDecoration.lerp] when interpolating two
   ///    [DashedDecoration]s or a [DashedDecoration] to or from null.
-  static DashedDecoration lerp(
-      DashedDecoration a, DashedDecoration b, double t) {
-    assert(t != null);
+  static DashedDecoration? lerp(
+      DashedDecoration? a, DashedDecoration? b, double t) {
     if (a == null && b == null) {
       return null;
     }
     if (a == null) {
-      return b.scale(t);
+      return b!.scale(t);
     }
     if (b == null) {
       return a.scale(1.0 - t);
@@ -251,7 +250,7 @@ class DashedDecoration extends Decoration {
       return b;
     }
     return DashedDecoration(
-      color: Color.lerp(a.color, b.color, t),
+      color: Color.lerp(a.color, b.color, t)!,
       image: t < 0.5 ? a.image : b.image,
 // TODO(ianh): cross-fade the image
       border: BoxBorder.lerp(a.border, b.border, t),
@@ -318,14 +317,14 @@ class DashedDecoration extends Decoration {
   }
 
   @override
-  bool hitTest(Size size, Offset position, {TextDirection textDirection}) {
+  bool hitTest(Size size, Offset position, {TextDirection? textDirection}) {
     assert(shape != null);
     assert((Offset.zero & size).contains(position));
     switch (shape) {
       case BoxShape.rectangle:
         if (borderRadius != null) {
           final RRect bounds =
-              borderRadius.resolve(textDirection).toRRect(Offset.zero & size);
+              borderRadius!.resolve(textDirection).toRRect(Offset.zero & size);
           return bounds.contains(position);
         }
         return true;
@@ -334,13 +333,13 @@ class DashedDecoration extends Decoration {
         final Offset center = size.center(Offset.zero);
         final double distance = (position - center).distance;
         return distance <= math.min(size.width, size.height) / 2.0;
+      default:
+        return false;
     }
-    assert(shape != null);
-    return null;
   }
 
   @override
-  _BoxDecorationPainter createBoxPainter([VoidCallback onChanged]) {
+  _BoxDecorationPainter createBoxPainter([VoidCallback? onChanged]) {
     assert(onChanged != null || image == null);
     return _BoxDecorationPainter(this, onChanged);
   }
@@ -348,17 +347,15 @@ class DashedDecoration extends Decoration {
 
 /// An object that paints a [DashedDecoration] into a canvas.
 class _BoxDecorationPainter extends BoxPainter {
-  _BoxDecorationPainter(this._decoration, VoidCallback onChanged)
-      : assert(_decoration != null),
-        super(onChanged);
+  _BoxDecorationPainter(this._decoration, VoidCallback? onChanged)
+      : super(onChanged);
 
   final DashedDecoration _decoration;
 
-  Paint _cachedBackgroundPaint;
-  Rect _rectForCachedBackgroundPaint;
+  Paint? _cachedBackgroundPaint;
+  Rect? _rectForCachedBackgroundPaint;
 
   Paint _getBackgroundPaint(Rect rect, TextDirection textDirection) {
-    assert(rect != null);
     assert(
         _decoration.gradient != null || _rectForCachedBackgroundPaint == null);
 
@@ -366,20 +363,21 @@ class _BoxDecorationPainter extends BoxPainter {
         (_decoration.gradient != null &&
             _rectForCachedBackgroundPaint != rect)) {
       final Paint paint = Paint();
-      if (_decoration.backgroundBlendMode != null)
-        paint.blendMode = _decoration.backgroundBlendMode;
+      if (_decoration.backgroundBlendMode != null) {
+        paint.blendMode = _decoration.backgroundBlendMode!;
+      }
       if (_decoration.color != null) {
-        paint.color = _decoration.color;
+        paint.color = _decoration.color!;
       }
       if (_decoration.gradient != null) {
-        paint.shader = _decoration.gradient
+        paint.shader = _decoration.gradient!
             .createShader(rect, textDirection: textDirection);
         _rectForCachedBackgroundPaint = rect;
       }
       _cachedBackgroundPaint = paint;
     }
 
-    return _cachedBackgroundPaint;
+    return _cachedBackgroundPaint!;
   }
 
   void _paintBox(
@@ -396,10 +394,12 @@ class _BoxDecorationPainter extends BoxPainter {
           canvas.drawRect(rect, paint);
         } else {
           canvas.drawRRect(
-              _decoration.borderRadius.resolve(textDirection).toRRect(rect),
+              _decoration.borderRadius!.resolve(textDirection).toRRect(rect),
               paint);
         }
         break;
+      default:
+        return;
     }
   }
 
@@ -407,7 +407,7 @@ class _BoxDecorationPainter extends BoxPainter {
     if (_decoration.boxShadow == null) {
       return;
     }
-    for (final BoxShadow boxShadow in _decoration.boxShadow) {
+    for (final BoxShadow boxShadow in _decoration.boxShadow!) {
       final Paint paint = boxShadow.toPaint();
       final Rect bounds =
           rect.shift(boxShadow.offset).inflate(boxShadow.spreadRadius);
@@ -417,33 +417,37 @@ class _BoxDecorationPainter extends BoxPainter {
 
   void _paintBackgroundColor(
       Canvas canvas, Rect rect, TextDirection textDirection) {
-    if (_decoration.color != null || _decoration.gradient != null)
+    if (_decoration.color != null || _decoration.gradient != null) {
       _paintBox(canvas, rect, _getBackgroundPaint(rect, textDirection),
           textDirection);
+    }
   }
 
-  DecorationImagePainter _imagePainter;
+  DecorationImagePainter? _imagePainter;
 
   void _paintBackgroundImage(
       Canvas canvas, Rect rect, ImageConfiguration configuration) {
     if (_decoration.image == null) {
       return;
     }
-    _imagePainter ??= _decoration.image.createPainter(onChanged);
-    Path clipPath;
+    _imagePainter ??= _decoration.image!.createPainter(onChanged!);
+    Path? clipPath;
     switch (_decoration.shape) {
       case BoxShape.circle:
         clipPath = Path()..addOval(rect);
         break;
       case BoxShape.rectangle:
-        if (_decoration.borderRadius != null)
+        if (_decoration.borderRadius != null) {
           clipPath = Path()
-            ..addRRect(_decoration.borderRadius
+            ..addRRect(_decoration.borderRadius!
                 .resolve(configuration.textDirection)
                 .toRRect(rect));
+        }
         break;
+      default:
+        return;
     }
-    _imagePainter.paint(canvas, rect, clipPath, configuration);
+    _imagePainter!.paint(canvas, rect, clipPath, configuration);
   }
 
   @override
@@ -455,25 +459,24 @@ class _BoxDecorationPainter extends BoxPainter {
   /// Paint the box decoration into the given location on the given canvas
   @override
   void paint(Canvas canvas, Offset offset, ImageConfiguration configuration) {
-    assert(configuration != null);
     assert(configuration.size != null);
-    final Rect rect = offset & configuration.size;
-    final TextDirection textDirection = configuration.textDirection;
+    final Rect rect = offset & configuration.size!;
+    final TextDirection textDirection = configuration.textDirection!;
     _paintShadows(canvas, rect, textDirection);
     _paintBackgroundColor(canvas, rect, textDirection);
     _paintBackgroundImage(canvas, rect, configuration);
-    if (_decoration.dawDashed != null && !_decoration.dawDashed) {
+    if (_decoration.dawDashed != null && !(_decoration.dawDashed ?? true)) {
       _decoration.border?.paint(
         canvas,
         rect,
-        shape: _decoration.shape,
+        shape: _decoration.shape!,
         borderRadius: _decoration.borderRadius as BorderRadius,
         textDirection: configuration.textDirection,
       );
       return;
     }
     final Paint dashedPaint = Paint()
-      ..color = _decoration.dashedColor
+      ..color = _decoration.dashedColor!
       ..strokeWidth = _decoration.strokeHeight
       ..style = PaintingStyle.stroke;
 
@@ -509,9 +512,9 @@ class _BoxDecorationPainter extends BoxPainter {
   }
 
   Path getDashedPath({
-    @required math.Point<double> a,
-    @required math.Point<double> b,
-    @required num gap,
+    required math.Point<double> a,
+    required math.Point<double> b,
+    required num gap,
   }) {
     final Size size = Size(b.x - a.x, b.y - a.y);
     final Path path = Path();
