@@ -1,17 +1,18 @@
 // 视觉功能
+import 'package:dokit/kit/common/basic_info.dart';
+import 'package:dokit/kit/kit.dart';
 import 'package:dokit/ui/resident_page.dart';
 import 'package:flutter/material.dart';
-
-import '../kit.dart';
-import 'basic_info.dart';
 
 abstract class CommonKit implements IKit {
   @override
   void tabAction() {
-    ResidentPage.residentPageKey.currentState.setState(() {
+    // ignore: invalid_use_of_protected_member
+    ResidentPage.residentPageKey.currentState?.setState(() {
       ResidentPage.tag = getKitName();
     });
   }
+
   Widget createDisplayPage();
 }
 
@@ -20,7 +21,7 @@ class CommonKitManager {
     CommonKitName.KIT_BASIC_INFO: BasicInfoKit(),
   };
 
-  CommonKitManager._privateConstructor() {}
+  CommonKitManager._privateConstructor();
 
   static final CommonKitManager _instance =
       CommonKitManager._privateConstructor();
@@ -29,14 +30,12 @@ class CommonKitManager {
 
   // 如果想要自定义实现，可以用这个方式进行覆盖。后续扩展入口
   void addKit(String tag, CommonKit kit) {
-    assert(tag != null && kit != null);
     kitMap[tag] = kit;
   }
 
-  T getKit<T extends CommonKit>(String name) {
-    assert(name != null);
+  T? getKit<T extends CommonKit?>(String name) {
     if (kitMap.containsKey(name)) {
-      return kitMap[name];
+      return kitMap[name] as T;
     }
     return null;
   }
